almost working to fix the finals
